requirements.txt
python-telegram-bot==20.5
aiohttp==3.9.3
beautifulsoup4==4.12.3
python-dotenv==1.0.0